scripts/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pyyaml>=6.0
Pillow>=10.0.0
inquirer>=3.1.0
colorama>=0.4.6
python-frontmatter>=1.0.0